I am retarded
Not all departments use the same short for all their classes as they do for their url....in fact,I think a minority do.

for example, http://contacts.ucalgary.ca/info/arts/courses/w19

The scraper ran that over,expecting 'ARTS' as the short for the classes.
Except there are ZERO classes with ARTS as their prefix....

It didn't shit a brick, it just acknowledged it found nothing, and then didn't save the nothing it found.

It now splits the pages on the html string 'unitis-courses-heading', and then passes those individual strips of the main to the course table parser.  before it does that,it uses a regex to locate the short being used on that table, and passes that along.

it now appears to correctly handle these cases.

.....wonders what other stupidity I am guilty of....

Signed-off-by: Athandreyal <darkprev@hotmail.com>